code review - delete file that use for testing - clear comment that useless
