+++ server/getData.ts
@@ -25,21 +25,17 @@ export interface getRecord {
   parentId: string;
 }
 
-
-
 export async function getDataByType(type: string): Promise<any[]> {
   try {
-
-
     const params: any = {
       TableName: "money",
-      IndexName: "TypeIndex", // your GSI name
+      IndexName: "TypeIndex", 
       KeyConditionExpression: "#t = :type",
       ExpressionAttributeNames: {
         "#t": "type"
       },
       ExpressionAttributeValues: {
-        ":type": type  // 直接用字符串
+        ":type": type  
       }
     };
     console.log(params);
@@ -53,11 +49,9 @@ export async function getDataByType(type: string): Promise<any[]> {
 
 export async function  getDataByTypeUser(type: string, user: string): Promise<any[]> {
   try {
-
-
     const params: any = {
       TableName: "money",
-      IndexName: "TypeUserIndex", // your GSI name
+      IndexName: "TypeUserIndex",
       KeyConditionExpression: "#type = :typeVal AND #user = :userVal",
       ExpressionAttributeNames: {
         "#type": "type",
@@ -70,7 +64,6 @@ export async function  getDataByTypeUser(type: string, user: string): Promise<an
     };
     console.log(params);
     const data = await docClient.send(new QueryCommand(params));
-
     return data.Items || [];
   } catch (error) {
     console.error("Error fetching money:", error);
